+++ covid19-etl/etl/ccmap.py
@@ -6,7 +6,7 @@ import re
 import requests
 
 from etl import base
-from helper.metadata_helper import MetadataHelper
+from utils.metadata_helper import MetadataHelper
 
 
 def format_location_submitter_id(in_json):
