+++ covid19-etl/etl/chi_nbhd.py
@@ -4,8 +4,8 @@ from contextlib import closing
 import requests
 
 from etl import base
-from helper.format_helper import derived_submitter_id, format_submitter_id
-from helper.metadata_helper import MetadataHelper
+from utils.format_helper import derived_submitter_id, format_submitter_id
+from utils.metadata_helper import MetadataHelper
 
 
 class CHI_NBHD(base.BaseETL):
