+++ covid19-etl/etl/coxray.py
@@ -2,9 +2,9 @@ import csv
 from pathlib import Path
 
 from etl import base
-from helper.file_helper import FileHelper
-from helper.format_helper import derived_submitter_id, format_submitter_id
-from helper.metadata_helper import MetadataHelper
+from utils.file_helper import FileHelper
+from utils.format_helper import derived_submitter_id, format_submitter_id
+from utils.metadata_helper import MetadataHelper
 
 COXRAY_DATA_PATH = "../data"
