+++ covid19-etl/etl/ds4c.py
@@ -2,8 +2,8 @@ import csv
 import os
 
 from etl import base
-from helper.format_helper import check_date_format
-from helper.metadata_helper import MetadataHelper
+from utils.format_helper import check_date_format
+from utils.metadata_helper import MetadataHelper
 
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
