+++ covid19-etl/etl/dsci.py
@@ -3,8 +3,8 @@ from datetime import datetime
 import os
 
 from etl import base
-from helper.format_helper import check_date_format
-from helper.metadata_helper import MetadataHelper
+from utils.format_helper import check_date_format
+from utils.metadata_helper import MetadataHelper
 
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
