+++ covid19-etl/etl/dsfsi.py
@@ -7,7 +7,7 @@ from contextlib import closing
 import requests
 
 from etl import base
-from helper.metadata_helper import MetadataHelper
+from utils.metadata_helper import MetadataHelper
 
 
 def format_subject_submitter_id(country, submitter_id):
