+++ covid19-etl/etl/idph_hospital.py
@@ -6,12 +6,12 @@ from contextlib import closing
 import requests
 
 from etl import base
-from helper.format_helper import (
+from utils.format_helper import (
     derived_submitter_id,
     format_submitter_id,
     idph_get_date,
 )
-from helper.metadata_helper import MetadataHelper
+from utils.metadata_helper import MetadataHelper
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
