+++ covid19-etl/etl/idph_zipcode.py
@@ -4,13 +4,13 @@ import re
 import requests
 
 from etl import base
-from helper.idph_helper import fields_mapping
-from helper.format_helper import (
+from utils.idph_helper import fields_mapping
+from utils.format_helper import (
     derived_submitter_id,
     format_submitter_id,
     idph_get_date,
 )
-from helper.metadata_helper import MetadataHelper
+from utils.metadata_helper import MetadataHelper
 
 
 class IDPH_ZIPCODE(base.BaseETL):
