+++ covid19-etl/etl/owid2.py
@@ -5,7 +5,7 @@ from contextlib import closing
 import requests
 
 from etl import base
-from helper.metadata_helper import MetadataHelper
+from utils.metadata_helper import MetadataHelper
 
 
 def format_location_submitter_id(country):
