+++ covid19-etl/etl/ssr.py
@@ -9,8 +9,8 @@ import os
 import xlrd
 
 from etl import base
-from helper.metadata_helper import MetadataHelper
-from helper.format_helper import format_submitter_id, derived_submitter_id
+from utils.metadata_helper import MetadataHelper
+from utils.format_helper import format_submitter_id, derived_submitter_id
 
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
